+++ docs/changelog.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Action and necessary logic to un-mitigate a hijack
 - Ability to reload targeted services based on what conf section changed
-- GET config endpoints to backend services
+- GET config endpoints to backend and monitor services
 
 ### Changed
 - Decoupled microservice architecture for backend and frontend
@@ -19,18 +19,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced bulk timer thread with separate process
 - Replaced autoignore checker thread with separate process
 - Replaced exabgp autoconf thread with separate process
+- Display info on new microservices, grouping backend/monitor
 - Upgraded requests to 2.25.1
+- Upgraded to new bgpstream containers (using also ris and routeviews live sources)
 
 ### Fixed
 - Discard messages with timestamps in future datetime (updated to artemis-utils==1.0.5)
 - Check for very old timestamps in DB update handling
 - Replaced sets that are not supported by shared manager with lists
+- Removed unneeded read locks
 
 ### Removed
 - TBD (removed a feature)
 
 ### Deprecated
-- TBD (soon-to-be removed feature)
+- 'caida' project in bgpstreamlive
 
 ### Security
 - TBD (addressing vulnerability)